Remove unnecessary classification map
