+++ src/EditorFeatures/Core/Shared/Utilities/ClassificationTypeMap.cs
@@ -3,20 +3,60 @@
 // See the LICENSE file in the project root for more information.
 
 using System;
+using System.Collections.Generic;
 using System.ComponentModel.Composition;
+using Microsoft.CodeAnalysis.Classification;
+using Microsoft.CodeAnalysis.ErrorReporting;
 using Microsoft.CodeAnalysis.Host.Mef;
 using Microsoft.VisualStudio.Text.Classification;
+using Roslyn.Utilities;
+using ReferenceEqualityComparer = Roslyn.Utilities.ReferenceEqualityComparer;
 
 namespace Microsoft.CodeAnalysis.Editor.Shared.Utilities;
 
-/// <summary>
-/// Note: we very intentionally place ourselves in the editors 'syntactic' bucket (and not 'semantic' bucket).  See
-/// comments on <see cref="AbstractClassificationTypeMap"/> for more details.
-/// </summary>
 [Export]
-[method: ImportingConstructor]
-[method: Obsolete(MefConstruction.ImportingConstructorMessage, error: true)]
-internal sealed class ClassificationTypeMap(IClassificationTypeRegistryService registryService)
-    : AbstractClassificationTypeMap(registryService, ClassificationLayer.Syntactic)
+internal sealed class ClassificationTypeMap : IClassificationTypeMap
 {
+    private readonly IClassificationTypeRegistryService _registryService;
+    private readonly Dictionary<string, IClassificationType> _identityMap;
+
+    [ImportingConstructor]
+    [Obsolete(MefConstruction.ImportingConstructorMessage, error: true)]
+    public ClassificationTypeMap(IClassificationTypeRegistryService registryService)
+    {
+        _registryService = registryService;
+
+        // Prepopulate the identity map with the constant string values from ClassificationTypeNames
+        var fields = typeof(ClassificationTypeNames).GetFields();
+        _identityMap = new Dictionary<string, IClassificationType>(fields.Length, ReferenceEqualityComparer.Instance);
+
+        foreach (var field in fields)
+        {
+            // The strings returned from reflection do not have reference-identity with the string constants used by
+            // the compiler. Fortunately, a call to string.Intern fixes them.
+            var rawValue = (string?)field.GetValue(null);
+            Contract.ThrowIfNull(rawValue);
+            var value = string.Intern(rawValue);
+
+            _identityMap.Add(value, registryService.GetClassificationType(ClassificationLayer.Semantic, value));
+        }
+    }
+
+    public IClassificationType GetClassificationType(string name)
+    {
+        var type = GetClassificationTypeWorker(name);
+        if (type == null)
+        {
+            FatalError.ReportAndCatch(new Exception($"classification type doesn't exist for {name}"));
+        }
+
+        return type ?? GetClassificationTypeWorker(ClassificationTypeNames.Text);
+    }
+
+    private IClassificationType GetClassificationTypeWorker(string name)
+    {
+        return _identityMap.TryGetValue(name, out var result)
+            ? result
+            : _registryService.GetClassificationType(ClassificationLayer.Semantic, name);
+    }
 }
